# docs/swagger.yaml
definitions:
...
  dtos.CategoryProducts:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  dtos.MessagesResponses:
    properties:
      message:
        type: string
    type: object
  dtos.Operate:
    properties:
      active:
        type: boolean
      close:
        type: string
      day_operational:
        items:
          enum:
          - senin
          - selasa
          - rabu
          - kamis
          - jum'at
          - sabtu
          - minggu
          type: string
        type: array
      open:
        type: string
    type: object
...
info:
  contact:
    email: cs@payuoge.com
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /groceries/operational:
    get:
      consumes:
      - application/json
      description: do Get all operational
      produces:
      - application/json
      responses:
        "200":
          description: the message successfully create
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetAllOperational access process
      tags:
      - groceries
    post:
      consumes:
      - application/json
      description: do create a operational
      parameters:
      - description: create a operational
        in: body
        name: operational
        required: true
        schema:
          $ref: '#/definitions/dtos.Operate'
      produces:
      - application/json
      responses:
        "200":
          description: the message successfully create
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: CreateOperational access process
      tags:
      - groceries
  /groceries/operational/{id}:
    delete:
      consumes:
      - application/json
      description: do delete operational
      parameters:
      - description: get id operational
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: delete operational successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: delete operational process
      tags:
      - groceries
    get:
      consumes:
      - application/json
      description: do Get ID operational
      parameters:
      - description: id operational
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - Bearer: []
      summary: GetIDOperational access process
      tags:
      - groceries
    put:
      consumes:
      - application/json
      description: do Update a operational
      parameters:
      - description: id operational
        in: path
        name: id
        required: true
        type: integer
      - description: update a operational
        in: body
        name: operational
        required: true
        schema:
          $ref: '#/definitions/dtos.Operate'
      responses:
        "200":
          description: the message successfully create
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: UpdateOperational access process
      tags:
      - groceries
...
  /products/category:
    get:
      consumes:
      - application/json
      description: do get all category
      produces:
      - application/json
      responses:
        "200":
          description: get all category
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookies not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetAll category process
      tags:
      - products
    post:
      consumes:
      - application/json
      description: do create category
      parameters:
      - description: create category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryProducts'
      produces:
      - application/json
      responses:
        "200":
          description: create category successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: create category process
      tags:
      - products
  /products/category/{id}:
    delete:
      consumes:
      - application/json
      description: do delete a category
      parameters:
      - description: delete size with category
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get all size
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: DeleteID category process
      tags:
      - products
    get:
      consumes:
      - application/json
      description: do get id category
      parameters:
      - description: get id category
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get category
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetID category process
      tags:
      - products
    put:
      consumes:
      - application/json
      description: do update category
      parameters:
      - description: access id category
        in: path
        name: id
        required: true
        type: integer
      - description: update category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryProducts'
      produces:
      - application/json
      responses:
        "200":
          description: update category successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: update category process
      tags:
      - products
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"